backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pymupdf>=1.23.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
